services/hornex-api/pyproject.toml
[tool.poetry]
name = "hornex-api"
version = "0.1.0"
description = ""
authors = ["Pedro <pedro357bm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
Markdown = "^3.4.4"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
coverage = "^7.3.0"
celery = "^5.3.4"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.2.0"
faker = "^20.0.0"
channels = "^4.0.0"
django-redis = "^5.4.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
channels-redis = "^4.1.0"
asgiref = "^3.7.2"
daphne = "^4.0.0"
whitenoise = "^6.6.0"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.1"
mock = "^5.1.0"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "DJ", # flake8-django
    "E", # pycodestyle Error
    "F", # Pyflakes
    "I", # isort
    "RUF", # Ruff-specific rules
    "UP", # pyupgrade
    "W", # pycodestyle Wrning
    "YTT", # flake8-2020
]
target-version = "py311"
ignore = [
  "E501",

  ## TODO Fix and enable these
  "DJ001",
  "DJ008",
  "DJ012",
  "RUF012",
  "UP008",
  ]

[tool.ruff.format]
quote-style = "double"
